Call supervisor/node via IPC instead of remote
