+++ src/shared/ipc.ts
@@ -27,6 +27,13 @@ export enum MainIpcListenChannels {
     ImportIdentity = "import-identity",
     PushSubscribe = "push-subscribe-country",
     PushUnsubscribe = "push-unsubscribe",
+    SupervisorConnect = "supervisor-connect",
+    SupervisorInstall = "supervisor-install",
+    SupervisorUpgrade = "supervisor-upgrade",
+    SupervisorDisconnect = "supervisor-disconnect",
+    KillGhosts = "kill-ghost",
+    StartNode = "start-node",
+    StopNode = "stop-node",
 }
 
 export interface IpcResponse {
